backend/requirements.txt
Flask==2.0.1
requests==2.31.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary
Flask-CORS

